requirements.txt
openai_whisper==20240930
